Report page with journeys

+++ app/reports/mutations/deleteReport.ts
@@ -6,9 +6,12 @@ const DeleteReport = z.object({
   id: z.number(),
 });
 
-export default resolver.pipe(resolver.zod(DeleteReport), resolver.authorize(), async ({ id }) => {
-  // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-  const report = await db.report.deleteMany({ where: { id } });
+export default resolver.pipe(
+  resolver.zod(DeleteReport),
+  resolver.authorize(),
+  async ({ id }, ctx) => {
+    const report = await db.report.deleteMany({ where: { id, userId: ctx.session.userId } });
 
-  return report;
-});
+    return report;
+  },
+);
